requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decouple==0.0.7
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-heroku==0.3.1
django-js-asset==1.2.2
docopt==0.6.2
filelock==3.0.12
fpdf==1.7.2
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.2
pipenv==2020.11.15
plotly==4.14.3
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==1.5
pytz==2021.1
requests==2.25.1
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.4
virtualenv==20.4.3
virtualenv-clone==0.5.4
whitenoise==5.2.0
